requirements.txt
# Contenido de requirements.txt (ACTUALIZADO)
flask
requests
gunicorn
google-generativeai
